backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.36
asyncpg==0.29.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
anthropic==0.40.0
pydantic-settings==2.6.1
alembic==1.14.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.3
httpx==0.28.1
